+++ AI_NETCORE_API/Domain/BusinessObject/BuyOffer.cs
@@ -6,11 +6,11 @@ namespace Domain.BusinessObject
 {
     public class BuyOffer : BaseOffer
     {
-        public BuyOffer(int id, int resourceId, int amount, DateTime date, bool isValid, decimal maxPrice) : base(id,resourceId, amount, date, isValid)
+        public BuyOffer(int id, int resourceId, int amount, DateTime date, bool isValid, double maxPrice) : base(id,resourceId, amount, date, isValid)
         {
             MaxPrice = maxPrice;
         }
 
-        public decimal MaxPrice { get; }
+        public double MaxPrice { get; }
     }
 }
